Introduce VWAP Data Point Interface
Point provides the data required to calculate a VWAP for a specific pair
from coinbase.
