+++ internal/storage/linked-list/vwap_linkedlist.go
@@ -7,6 +7,11 @@ import (
 	"sync"
 )
 
+/**
+ * @author Mohamed-Aly Bou-Hanane
+ * © 2022
+ */
+
 // vwapLinkedList represents a doubly linked list as a queue of DataPoints.
 // Manipulation with Linked List is faster than Array List because it uses a doubly linked list, so no bit shifting is required in memory.
 //Every time a new data point is added to the queue and saved for each trading pair, the VWAP computation is updated accordingly.
