+++ internal/storage/linked-list/vwap_linkedlist_test.go
@@ -10,6 +10,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+/**
+ * @author Mohamed-Aly Bou-Hanane
+ * © 2022
+ */
+
 var points = map[string]storage.Point{
 	"1": storage.NewPoint(1, 1, "TradingPair1"),
 	"2": storage.NewPoint(2, 2, "TradingPair2"),
